LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A5.7_T1-expected.txt

S7.9_A5.7_T1

PASS ReferenceError: Prefix ++ operator applied to value that is not a reference.

TEST COMPLETE
